Harden net client and snapshot against malformed/hostile input
Client (net_client.c):
- HandleProfile read a 4-byte playerId appended after the name but its
  length guard still only required the pre-playerId body, so a short frame
  passed the check and ReadU32(payload + 20 + NAME_LEN) read past the end.
  Require the four extra bytes.
- Net_GetRates used experience == 0 as "no rates arrived yet", which
  silently discards a server that legitimately runs a 0x experience
  ruleset (all six rates fall back to the identity). Track arrival with an
  explicit hasRates flag instead.
- Net_SendKeys: guard the batch write and loop to the KEYS_BATCH_FRAMES
  constant so every body[] index is provably in bounds (silences a
  stringop-overflow warning and removes the last-slot edge case).

Server (world.rs): pose.x/y are client-controlled i16, so a hostile client
can plant -32768. The snapshot nearest-8 sort and the update_pose step
check both subtracted in i16 before widening, which overflows (a panic
under overflow-checks, a wrap in release) or hits abs(i16::MIN). Widen to
i32 before subtracting so any reported pose is safe.

+++ server/server/src/world.rs
@@ -242,8 +242,10 @@ impl World {
             return Some(pose);
         }
 
-        let dx = (pose.x - p.pose.x).abs();
-        let dy = (pose.y - p.pose.y).abs();
+        // Widen before subtracting/abs: both operands are client-controlled i16, and abs() on
+        // i16::MIN panics just as the subtraction can overflow. i32 makes every reported pose safe.
+        let dx = (pose.x as i32 - p.pose.x as i32).abs();
+        let dy = (pose.y as i32 - p.pose.y as i32).abs();
 
         // Standing still, turning, or a single step along one axis. Diagonals are not a
         // thing the player avatar can do.
@@ -307,8 +309,11 @@ impl World {
 
         if visible.len() > MAX_VISIBLE_PLAYERS {
             visible.sort_by_key(|p| {
-                let dx = (p.pose.x - from.x) as i32;
-                let dy = (p.pose.y - from.y) as i32;
+                // Widen before subtracting: pose.x/y are client-controlled i16 and a hostile
+                // client can plant -32768, so an i16 subtraction here would overflow (panic under
+                // overflow-checks, wrap in release) on a value another player never chose.
+                let dx = p.pose.x as i32 - from.x as i32;
+                let dy = p.pose.y as i32 - from.y as i32;
                 dx * dx + dy * dy
             });
             visible.truncate(MAX_VISIBLE_PLAYERS);
